+++ overlay/var/www/peterlyons.com/problog/wp-admin/users.php
@@ -7,16 +7,27 @@
  */
 
 /** WordPress Administration Bootstrap */
-require_once('admin.php');
+require_once('./admin.php');
 
 /** WordPress Registration API */
 require_once( ABSPATH . WPINC . '/registration.php');
 
-if ( !current_user_can('edit_users') )
+if ( !current_user_can('list_users') )
 	wp_die(__('Cheatin&#8217; uh?'));
 
 $title = __('Users');
 $parent_file = 'users.php';
+
+// contextual help - choose Help on the top right of admin panel to preview this.
+add_contextual_help($current_screen,
+    '<p>' . __('This screen lists all the existing users for your site. Each user has one of five defined roles as set by the site admin: Site Administrator, Editor, Author, Contributor, or Subscriber. Users with roles other than Administrator will see fewer options when they are logged in, based on their role.') . '</p>' .
+    '<p>' . __('You can customize the display of information on this screen as you can on other screens, by using the Screen Options tab and the on-screen filters.') . '</p>' .
+    '<p>' . __('To add a new user for your site, click the Add New button at the top of the screen or Add New in the Users menu section.') . '</p>' .
+    '<p><strong>' . __('For more information:') . '</strong></p>' .
+    '<p>' . __('<a href="http://codex.wordpress.org/Users_Authors_and_Users_SubPanel" target="_blank">Documentation on Authors and Users</a>') . '</p>' .
+    '<p>' . __('<a href="http://codex.wordpress.org/Roles_and_Capabilities" target="_blank">Roles and Capabilities Descriptions</a>') . '</p>' .
+    '<p>' . __('<a href="http://wordpress.org/support/" target="_blank">Support Forums</a>') . '</p>'
+);
 
 $update = $doaction = '';
 if ( isset($_REQUEST['action']) )
@@ -43,25 +54,31 @@ switch ($doaction) {
 case 'promote':
 	check_admin_referer('bulk-users');
 
-	if (empty($_REQUEST['users'])) {
+	if ( empty($_REQUEST['users']) ) {
 		wp_redirect($redirect);
 		exit();
 	}
 
 	$editable_roles = get_editable_roles();
-	if (!$editable_roles[$_REQUEST['new_role']])
+	if ( empty( $editable_roles[$_REQUEST['new_role']] ) )
 		wp_die(__('You can&#8217;t give users that role.'));
 
 	$userids = $_REQUEST['users'];
 	$update = 'promote';
-	foreach($userids as $id) {
-		if ( ! current_user_can('edit_user', $id) )
+	foreach ( $userids as $id ) {
+		$id = (int) $id;
+
+		if ( ! current_user_can('promote_user', $id) )
 			wp_die(__('You can&#8217;t edit that user.'));
-		// The new role of the current user must also have edit_users caps
-		if($id == $current_user->ID && !$wp_roles->role_objects[$_REQUEST['new_role']]->has_cap('edit_users')) {
+		// The new role of the current user must also have promote_users caps
+		if ( $id == $current_user->ID && !$wp_roles->role_objects[$_REQUEST['new_role']]->has_cap('promote_users') ) {
 			$update = 'err_admin_role';
 			continue;
 		}
+
+		// If the user doesn't already belong to the blog, bail.
+		if ( is_multisite() && !is_user_member_of_blog( $id ) )
+			wp_die(__('Cheatin&#8217; uh?'));
 
 		$user = new WP_User($id);
 		$user->set_role($_REQUEST['new_role']);
@@ -73,6 +90,8 @@ case 'promote':
 break;
 
 case 'dodelete':
+	if ( is_multisite() )
+		wp_die( __('User deletion is not allowed from this screen.') );
 
 	check_admin_referer('delete-users');
 
@@ -81,7 +100,7 @@ case 'dodelete':
 		exit();
 	}
 
-	if ( !current_user_can('delete_users') )
+	if ( ! current_user_can( 'delete_users' ) )
 		wp_die(__('You can&#8217;t delete users.'));
 
 	$userids = $_REQUEST['users'];
@@ -89,19 +108,23 @@ case 'dodelete':
 	$delete_count = 0;
 
 	foreach ( (array) $userids as $id) {
-		if ( ! current_user_can('delete_user', $id) )
-			wp_die(__('You can&#8217;t delete that user.'));
+		$id = (int) $id;
 
-		if($id == $current_user->ID) {
+		if ( ! current_user_can( 'delete_user', $id ) )
+			wp_die(__( 'You can&#8217;t delete that user.' ) );
+
+		if ( $id == $current_user->ID ) {
 			$update = 'err_admin_del';
 			continue;
 		}
-		switch($_REQUEST['delete_option']) {
+		switch ( $_REQUEST['delete_option'] ) {
 		case 'delete':
-			wp_delete_user($id);
+			if ( current_user_can('delete_user', $id) )
+				wp_delete_user($id);
 			break;
 		case 'reassign':
-			wp_delete_user($id, $_REQUEST['reassign_user']);
+			if ( current_user_can('delete_user', $id) )
+				wp_delete_user($id, $_REQUEST['reassign_user']);
 			break;
 		}
 		++$delete_count;
@@ -114,6 +137,8 @@ case 'dodelete':
 break;
 
 case 'delete':
+	if ( is_multisite() )
+		wp_die( __('User deletion is not allowed from this screen.') );
 
 	check_admin_referer('bulk-users');
 
@@ -122,8 +147,8 @@ case 'delete':
 		exit();
 	}
 
-	if ( !current_user_can('delete_users') )
-		$errors = new WP_Error('edit_users', __('You can&#8217;t delete users.'));
+	if ( ! current_user_can( 'delete_users' ) )
+		$errors = new WP_Error( 'edit_users', __( 'You can&#8217;t delete users.' ) );
 
 	if ( empty($_REQUEST['users']) )
 		$userids = array(intval($_REQUEST['user']));
@@ -153,7 +178,13 @@ case 'delete':
 			$go_delete = true;
 		}
 	}
-	$all_logins = $wpdb->get_results("SELECT ID, user_login FROM $wpdb->users ORDER BY user_login");
+	// @todo Delete is always for !is_multisite(). Use API.
+	if ( !is_multisite() ) {
+		$all_logins = $wpdb->get_results("SELECT ID, user_login FROM $wpdb->users ORDER BY user_login");
+	} else {
+		// WPMU only searches users of current blog
+		$all_logins = $wpdb->get_results("SELECT ID, user_login FROM $wpdb->users, $wpdb->usermeta WHERE $wpdb->users.ID = $wpdb->usermeta.user_id AND meta_key = '".$wpdb->prefix."capabilities' ORDER BY user_login");
+	}
 	$user_dropdown = '<select name="reassign_user">';
 	foreach ( (array) $all_logins as $login )
 		if ( $login->ID == $current_user->ID || !in_array($login->ID, $userids) )
@@ -180,6 +211,94 @@ case 'delete':
 
 break;
 
+case 'doremove':
+	check_admin_referer('remove-users');
+
+	if ( empty($_REQUEST['users']) ) {
+		wp_redirect($redirect);
+		exit;
+	}
+
+	if ( !current_user_can('remove_users')  )
+		die(__('You can&#8217;t remove users.'));
+
+	$userids = $_REQUEST['users'];
+
+	$update = 'remove';
+ 	foreach ( $userids as $id ) {
+		$id = (int) $id;
+		if ( $id == $current_user->id && !is_super_admin() ) {
+			$update = 'err_admin_remove';
+			continue;
+		}
+		if ( !current_user_can('remove_user', $id) ) {
+			$update = 'err_admin_remove';
+			continue;
+		}
+		remove_user_from_blog($id, $blog_id);
+	}
+
+	$redirect = add_query_arg( array('update' => $update), $redirect);
+	wp_redirect($redirect);
+	exit;
+
+break;
+
+case 'remove':
+
+	check_admin_referer('bulk-users');
+
+	if ( empty($_REQUEST['users']) && empty($_REQUEST['user']) ) {
+		wp_redirect($redirect);
+		exit();
+	}
+
+	if ( !current_user_can('remove_users') )
+		$error = new WP_Error('edit_users', __('You can&#8217;t remove users.'));
+
+	if ( empty($_REQUEST['users']) )
+		$userids = array(intval($_REQUEST['user']));
+	else
+		$userids = $_REQUEST['users'];
+
+	include ('admin-header.php');
+?>
+<form action="" method="post" name="updateusers" id="updateusers">
+<?php wp_nonce_field('remove-users') ?>
+<?php echo $referer; ?>
+
+<div class="wrap">
+<?php screen_icon(); ?>
+<h2><?php _e('Remove Users from Site'); ?></h2>
+<p><?php _e('You have specified these users for removal:'); ?></p>
+<ul>
+<?php
+	$go_remove = false;
+ 	foreach ( $userids as $id ) {
+		$id = (int) $id;
+ 		$user = new WP_User($id);
+		if ( $id == $current_user->id && !is_super_admin() ) {
+			echo "<li>" . sprintf(__('ID #%1s: %2s <strong>The current user will not be removed.</strong>'), $id, $user->user_login) . "</li>\n";
+		} elseif ( !current_user_can('remove_user', $id) ) {
+			echo "<li>" . sprintf(__('ID #%1s: %2s <strong>You don\'t have permission to remove this user.</strong>'), $id, $user->user_login) . "</li>\n";
+		} else {
+			echo "<li><input type=\"hidden\" name=\"users[]\" value=\"{$id}\" />" . sprintf(__('ID #%1s: %2s'), $id, $user->user_login) . "</li>\n";
+			$go_remove = true;
+		}
+ 	}
+ 	?>
+<?php if ( $go_remove ) : ?>
+		<input type="hidden" name="action" value="doremove" />
+		<p class="submit"><input type="submit" name="submit" value="<?php esc_attr_e('Confirm Removal'); ?>" class="button-secondary" /></p>
+<?php else : ?>
+	<p><?php _e('There are no valid users selected for removal.'); ?></p>
+<?php endif; ?>
+</div>
+</form>
+<?php
+
+break;
+
 default:
 
 	if ( !empty($_GET['_wp_http_referer']) ) {
@@ -187,14 +306,20 @@ default:
 		exit;
 	}
 
-	include('admin-header.php');
+	include('./admin-header.php');
 
 	$usersearch = isset($_GET['usersearch']) ? $_GET['usersearch'] : null;
 	$userspage = isset($_GET['userspage']) ? $_GET['userspage'] : null;
 	$role = isset($_GET['role']) ? $_GET['role'] : null;
 
-	// Query the users
+	// Query the user IDs for this page
 	$wp_user_search = new WP_User_Search($usersearch, $userspage, $role);
+
+	// Query the post counts for this page
+	$post_counts = count_many_users_posts($wp_user_search->get_results());
+
+	// Query the users for this page
+	cache_users($wp_user_search->get_results());
 
 	$messages = array();
 	if ( isset($_GET['update']) ) :
@@ -202,21 +327,28 @@ default:
 		case 'del':
 		case 'del_many':
 			$delete_count = isset($_GET['delete_count']) ? (int) $_GET['delete_count'] : 0;
-			$messages[] = '<div id="message" class="updated fade"><p>' . sprintf(_n('%s user deleted', '%s users deleted', $delete_count), $delete_count) . '</p></div>';
+			$messages[] = '<div id="message" class="updated"><p>' . sprintf(_n('%s user deleted', '%s users deleted', $delete_count), $delete_count) . '</p></div>';
 			break;
 		case 'add':
-			$messages[] = '<div id="message" class="updated fade"><p>' . __('New user created.') . '</p></div>';
+			$messages[] = '<div id="message" class="updated"><p>' . __('New user created.') . '</p></div>';
 			break;
 		case 'promote':
-			$messages[] = '<div id="message" class="updated fade"><p>' . __('Changed roles.') . '</p></div>';
+			$messages[] = '<div id="message" class="updated"><p>' . __('Changed roles.') . '</p></div>';
 			break;
 		case 'err_admin_role':
 			$messages[] = '<div id="message" class="error"><p>' . __('The current user&#8217;s role must have user editing capabilities.') . '</p></div>';
-			$messages[] = '<div id="message" class="updated fade"><p>' . __('Other user roles have been changed.') . '</p></div>';
+			$messages[] = '<div id="message" class="updated"><p>' . __('Other user roles have been changed.') . '</p></div>';
 			break;
 		case 'err_admin_del':
 			$messages[] = '<div id="message" class="error"><p>' . __('You can&#8217;t delete the current user.') . '</p></div>';
-			$messages[] = '<div id="message" class="updated fade"><p>' . __('Other users have been deleted.') . '</p></div>';
+			$messages[] = '<div id="message" class="updated"><p>' . __('Other users have been deleted.') . '</p></div>';
+			break;
+		case 'remove':
+			$messages[] = '<div id="message" class="updated fade"><p>' . __('User removed from this site.') . '</p></div>';
+			break;
+		case 'err_admin_remove':
+			$messages[] = '<div id="message" class="error"><p>' . __("You can't remove the current user.") . '</p></div>';
+			$messages[] = '<div id="message" class="updated fade"><p>' . __('Other users have been removed.') . '</p></div>';
 			break;
 		}
 	endif; ?>
@@ -239,7 +371,7 @@ if ( ! empty($messages) ) {
 
 <div class="wrap">
 <?php screen_icon(); ?>
-<h2><?php echo esc_html( $title ); ?>  <a href="user-new.php" class="button add-new-h2"><?php echo esc_html_x('Add New', 'user'); ?></a> <?php
+<h2><?php echo esc_html( $title ); if ( current_user_can( 'create_users' ) ) { ?>  <a href="user-new.php" class="button add-new-h2"><?php echo esc_html_x('Add New', 'user'); ?></a><?php }
 if ( isset($_GET['usersearch']) && $_GET['usersearch'] )
 	printf( '<span class="subtitle">' . __('Search results for &#8220;%s&#8221;') . '</span>', esc_html( $_GET['usersearch'] ) ); ?>
 </h2>
@@ -248,22 +380,14 @@ if ( isset($_GET['usersearch']) && $_GET['usersearch'] )
 <form id="list-filter" action="" method="get">
 <ul class="subsubsub">
 <?php
-$role_links = array();
-$avail_roles = array();
-$users_of_blog = get_users_of_blog();
-$total_users = count( $users_of_blog );
-foreach ( (array) $users_of_blog as $b_user ) {
-	$b_roles = unserialize($b_user->meta_value);
-	foreach ( (array) $b_roles as $b_role => $val ) {
-		if ( !isset($avail_roles[$b_role]) )
-			$avail_roles[$b_role] = 0;
-		$avail_roles[$b_role]++;
-	}
-}
+$users_of_blog = count_users();
+$total_users = $users_of_blog['total_users'];
+$avail_roles =& $users_of_blog['avail_roles'];
 unset($users_of_blog);
 
 $current_role = false;
 $class = empty($role) ? ' class="current"' : '';
+$role_links = array();
 $role_links[] = "<li><a href='users.php'$class>" . sprintf( _nx( 'All <span class="count">(%s)</span>', 'All <span class="count">(%s)</span>', $total_users, 'users' ), number_format_i18n( $total_users ) ) . '</a>';
 foreach ( $wp_roles->get_names() as $this_role => $name ) {
 	if ( !isset($avail_roles[$this_role]) )
@@ -306,7 +430,11 @@ unset($role_links);
 <div class="alignleft actions">
 <select name="action">
 <option value="" selected="selected"><?php _e('Bulk Actions'); ?></option>
+<?php if ( !is_multisite() && current_user_can('delete_users') ) { ?>
 <option value="delete"><?php _e('Delete'); ?></option>
+<?php } else { ?>
+<option value="remove"><?php _e('Remove'); ?></option>
+<?php } ?>
 </select>
 <input type="submit" value="<?php esc_attr_e('Apply'); ?>" name="doaction" id="doaction" class="button-secondary action" />
 <label class="screen-reader-text" for="new_role"><?php _e('Change role to&hellip;') ?></label><select name="new_role" id="new_role"><option value=''><?php _e('Change role to&hellip;') ?></option><?php wp_dropdown_roles(); ?></select>
@@ -356,8 +484,11 @@ foreach ( $wp_user_search->get_results() as $userid ) {
 	$roles = $user_object->roles;
 	$role = array_shift($roles);
 
+	if ( is_multisite() && empty( $role ) )
+		continue;
+
 	$style = ( ' class="alternate"' == $style ) ? '' : ' class="alternate"';
-	echo "\n\t" . user_row($user_object, $style, $role);
+	echo "\n\t", user_row( $user_object, $style, $role, $post_counts[ $userid ] );
 }
 ?>
 </tbody>
@@ -372,8 +503,11 @@ foreach ( $wp_user_search->get_results() as $userid ) {
 <div class="alignleft actions">
 <select name="action2">
 <option value="" selected="selected"><?php _e('Bulk Actions'); ?></option>
+<?php if ( !is_multisite() && current_user_can('delete_users') ) { ?>
 <option value="delete"><?php _e('Delete'); ?></option>
-</select>
+<?php } else { ?>
+<option value="remove"><?php _e('Remove'); ?></option>
+<?php } ?></select>
 <input type="submit" value="<?php esc_attr_e('Apply'); ?>" name="doaction2" id="doaction2" class="button-secondary action" />
 </div>
 
@@ -385,11 +519,21 @@ foreach ( $wp_user_search->get_results() as $userid ) {
 </form>
 </div>
 
+<?php
+if ( is_multisite() ) {
+	foreach ( array('user_login' => 'user_login', 'first_name' => 'user_firstname', 'last_name' => 'user_lastname', 'email' => 'user_email', 'url' => 'user_uri', 'role' => 'user_role') as $formpost => $var ) {
+		$var = 'new_' . $var;
+		$$var = isset($_REQUEST[$formpost]) ? esc_attr(stripslashes($_REQUEST[$formpost])) : '';
+	}
+	unset($name);
+}
+?>
+
 <br class="clear" />
 <?php
 break;
 
 } // end of the $doaction switch
 
-include('admin-footer.php');
+include('./admin-footer.php');
 ?>
